requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
opensearch-py==2.4.0
aiohttp==3.9.5